% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFuns.R
\name{branin_with_grad}
\alias{branin_with_grad}
\title{Branin Function with Gradient}
\usage{
branin_with_grad(x)
}
\arguments{
\item{x}{A vector with length \eqn{d = 2}.}
}
\value{
A list with two numeric elements named \code{objective}
    and \code{gradient} corresponding to the function value
    (length 1) and its gradient (length 2).
}
\description{
Branin function with its exact gradient returned into
    a list.
}
\examples{
X <- matrix(runif(100), ncol = 2, dimnames = list(NULL, c("x1", "x2")))
Obj <- apply(X, 1, function(x) branin_with_grad(x)$objective)
ObjDK <- apply(X, 1, DiceKriging::branin)
max(abs(Obj - ObjDK))
Grad <- t(apply(X, 1, function(x) branin_with_grad(x)$gradient))
GradNum <- t(apply(X, 1, function(x) grad(DiceKriging::branin, x)))
max(abs(Grad - GradNum))

}
\seealso{
\code{\link[DiceKriging]{branin}}.
}
