% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{contours}
\alias{contours}
\title{Contours of a Function Related to a Kriging Model object
    with Two Inputs}
\usage{
contours(
  object,
  which = "mean",
  other = NULL,
  lower = c(0, 0),
  upper = c(1, 1),
  nGrid = 50,
  grad = FALSE,
  otherGrad = NULL,
  whereGrad = "levels",
  ...
)
}
\arguments{
\item{object}{An object with class \code{"km"} or equivalent
with two inputs.}

\item{which}{Character vector. Name of the kriging
function(s)/statistic(s) to be shown. Can contain
\code{"mean"}, \code{"trend"}, \code{"sd"}, \code{"sd2"} or
\code{"var"}. Can have length 0 as well, in order to display
the contours of a function having or not a formal argument
\code{object}. This function is then provided by using
\code{other}.}

\item{other}{Character. Name of a function. This can be a function
to be compared to the stat(s) given in \code{which} (usually
the kriging mean). If \code{which} has length zero we can also
specify a gradient function. The function must be a function
of one vector argument. It optionally can have one more
argument with name \code{"model"} or \code{"object"} in which
case \code{object} will be passed will be passed to this
argument as is required to diplay a one-point Bayesian
optimisation criterion (EI, AEI, ...), see \bold{Examples}.}

\item{lower, upper}{Numeric vectors with length 2 definign the
bounds of the rectangular region for the contours.}

\item{nGrid}{Integer. Number of grid points. Can be of length 2 if
different grids are to be used for the two dimensions.}

\item{grad}{Logical. If \code{TRUE} the gradient of the kriging
function/statistic will be shown using arrows. This will only
be possible when \emph{only one} function is used among
\code{"mean"} \code{"trend"} \code{"sd"} or \code{"sd2"}, or
when a function is given in \code{other}. In the second case
the gradient function must be given in \code{otherGrad}.}

\item{otherGrad}{Character. Name of the gradient function of the
function given in \code{other}. This will be ignored if
\code{other} is not provided or if \code{grad} is
\code{FALSE}. \bold{Not implemented yet}.}

\item{whereGrad}{Either a character with length one or a data
frame or matrix with two columns. This defines the points at
which the gradient will be evaluated and plotted as a small
arrow in a vector-field fashion. The default is
\code{"levels"} indicating that the gradient is evaluated
along some level curves (not those used to define the filled
contours). Another possible value is \code{"grid"} Finally if
a matrix or data frame is provided, the gradient will only be
evaluated and plotted at the specified points.}

\item{...}{Not used yet.}
}
\value{
A graphical object inheriting from \code{"ggplot"}.
}
\description{
Contours of a kriging function or statistic related
    to a kriging model object with two inputs. This can be the
    kriging mean, the kriging (conditional) standard deviation and
    more.
}
\details{
When several functions are given, they will be displayed
    on different facets of the plot. These functions should then
    have the same dimension because \emph{only one colour scale}
    is used.  For instance it makes no sense to give the mean and
    the variance which do not have the same dimension, nor even
    the mean and the standard deviation because these two
    functions may have different order of magnitude. We can
    usefully want the true function and the kriging mean, or the
    kriging mean and the kriging trend.
}
\examples{
set.seed(1521)
n <- 9
design <- matrix(runif(n * 2), ncol = 2,
                 dimnames = list(NULL, c("x1", "x2")))
design <- data.frame(design)
y <- apply(design, 1, branin)
model <- km(~1, design = design, response = y)

## Contours only
## =============
contours(model) + ggtitle("kriging mean (default in 'contours')")
contours(model, other = "branin") + ggtitle("'branin' and kriging mean")
contours(model, which = c("mean", "trend")) + ggtitle("Kriging mean and trend")
contours(model, which = "sd") + ggtitle("Kriging sd")

## Display gradients
## =================
contours(model, which = "mean", grad = TRUE) +
     ggtitle("Kriging mean, gradients at contours")
contours(model, which = "sd", grad = TRUE) +
    ggtitle("Kriging sd, gradients at contours")
contours(model, which = "sd", grad = TRUE, whereGrad = "grid") +
    ggtitle("Kriging sd, gradients at grid points")

## more involved examples
## ======================
\dontrun{
braninG <- function(x) braninGrad(x)$gradient
contours(model, which = character(0), grad = TRUE,
         other = "branin", otherGrad = "braninG", whereGrad = "grid") +
    ggtitle("Branin function and its gradient")
contours(model, which = character(0), grad = TRUE,
         other = "EI", otherGrad = "EI.grad", whereGrad = "grid") +
    ggtitle("Expected Improvement and its gradient")
}
}
