% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{dcrossprod}
\alias{dcrossprod}
\title{Diagonal of the Cross-Product of a Matrix}
\usage{
dcrossprod(X)
}
\arguments{
\item{X}{A numeric matrix.}
}
\value{
The diagonal vector of \code{crossprod(X)}.
}
\description{
Return the diagonal of the cross-product
\eqn{\mathbf{X}^\top\mathbf{X}}{t(X) \%*\% X} for a given matrix
\eqn{\mathbf{X}}{X}. As is well-known the cross-product is
computed by the single argument of \code{crossprod}; however we
often have to compute \emph{only the diagonal} of the
cross-product. For instance, the variance of a kriging predition
may have to be computed without formin the covariance.
}
\details{
Several methods can be used to compute the result in R (using
\code{apply}, ...). To our best knowledge the method used here is
the most efficient in general framework.
}
\author{
Thanks to Clément Walter for comparisons and the suggested
    code.
}
