% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolkaPack.R
\docType{package}
\name{dolka-package}
\alias{dolka-package}
\title{Utility Functions for Bayesian Optimization}
\description{
Transitional package for Bayes Optimization using
    either \pkg{DiceKriging} or \pkg{rlikriging} for the kriging
    steps.
}
\note{
The package is \emph{transitional}, meaning that it may
    exist only for a quite short period. Hopefully the implemented
    functions may continue to be available in other R packages.
}
\section{Authors}{

    \itemize{
   
     \item{

        Most of the functions devoted to Kriging are to be
        credited to the authors of the \pkg{DiceKriging} package:
        Oliver Roustant, David Ginsbourger and Yves Deville, or to
        its contributors: Clement Chevalier and Yann Richet.

     }

    \item {

        Most of the functions devoted to Bayesian Optimization are
        to be credited to the authors of the \pkg{DiceOptim}:
        Victor Picheny, David Ginsbourger and Olivier Roustant or
        to contributors: Michael Binois, Sebastien Marmin and
        Tobias Wagner.

     }
   
     \item {  

        Most of the functions devoted to kriging inversion are to
        be credited to the authors of the \pkg{KrigInv} package:
        Clement Chevalier, Victor Picheny, David Ginsbourger and
        Dario Azzimonti.

     }

     Yves Deville performed some refactoring of the functions and
     of their examples. Note that this can result in new errors. In
     case of doubt, use the original packages.

 }
}

