% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EI.R
\name{EI.grad}
\alias{EI.grad}
\title{Analytical gradient of the Expected Improvement criterion}
\usage{
EI.grad(
  x,
  model,
  plugin = NULL,
  type = c("UK", "SK"),
  minimization = TRUE,
  envir = NULL,
  proxy = FALSE
)
}
\arguments{
\item{x}{A vector representing the input for which one wishes to
calculate \code{\link{EI}}.}

\item{model}{An object of class \code{\link[DiceKriging]{km}}.}

\item{plugin}{Optional scalar: if provided, it replaces the
minimum of the current observations.}

\item{type}{Character \code{"UK"} (default) or \code{"SK"},
depending whether uncertainty related to trend estimation has
to be taken into account.}

\item{minimization}{Logical specifying if EI is used in
minimiziation or in maximization.}

\item{envir}{Optional environment specifying where to get
intermediate values calculated in \code{\link{EI}}.}

\item{proxy}{Optional logical. If \code{TRUE}, EI is replaced by the
kriging mean (to minimize).}
}
\value{
The gradient of the expected improvement criterion with
    respect to \eqn{x}.  Returns 0 at design points (where the
    gradient does not exist).
}
\description{
Computes the gradient of the Expected Improvement at
    the current location.  The current minimum of the observations
    can be replaced by an arbitrary value (plugin), which is
    useful in particular in noisy frameworks.
}
\section{Caution}{
 XXX the code uses \code{model@covariance@s2}
    which may not exist if some other class of object or even of
    covariance is used.
}

\examples{
set.seed(123)
## =========================================================================
## 	EI Surface Associated with an Ordinary Kriging Model for the Branin    
##  Function Known at a 9-Points Factorial Design  
## ========================================================================

## a 9-points factorial design, and the corresponding response
## ===========================================================
d <- 2; n <- 9
design.fact <- expand.grid(x1 = seq(0, 1, length=3), x2 = seq(0, 1, length=3))
y <- apply(design.fact, 1, branin) 

## model fit
## =========
fit1 <- km(~1, design = design.fact, response = y, covtype = "gauss",
           control = list(pop.size = 50, trace = FALSE), parinit = c(0.5, 0.5))

## graphics
## ========
n.grid <- 9  # Increase to 50 for a nicer picture
x.grid <- y.grid <- seq(0, 1, length = n.grid)
design.grid <- expand.grid(x1 = x.grid, x2 = y.grid)
EI.grid <- apply(design.grid, MARGIN = 1, FUN = EI, model = fit1)

z.grid <- matrix(EI.grid, n.grid, n.grid)

contour(x.grid, y.grid, z.grid, 20)
title("Expected Improvement for the Branin function known at 9 points")
points(design.fact[ , 1], design.fact[ , 2], pch = 17, col = "blue")

# graphics
n.gridx <- 5  # increase to 15 for nicer picture
n.gridy <- 5  # increase to 15 for nicer picture
x.grid2 <- seq(0, 1, length = n.gridx) 
y.grid2 <- seq(0, 1, length = n.gridy) 
design.grid2 <- expand.grid(x1 = x.grid2, x2 = y.grid2)

EI.envir <- new.env()	
environment(EI) <- environment(EI.grad) <- EI.envir 

for(i in 1:nrow(design.grid2)) {
	x <- design.grid2[i,  ]
	ei <- EI(x, model = fit1, envir = EI.envir)
	eigrad <- EI.grad(x , model = fit1, envir = EI.envir)
	if (!(is.null(ei))) {
	suppressWarnings(arrows(x$x1, x$x2,
	   x$x1 + eigrad[1] * 2.2 * 1e-4,
       x$x2 + eigrad[2] * 2.2 * 1e-4, 
	   length = 0.04, code = 2, col = "orange", lwd = 2))
	}
}

}
\references{
D. Ginsbourger (2009), \emph{Multiples metamodeles pour
l'approximation et l'optimisation de fonctions numeriques
multivariables}, Ph.D. thesis, Ecole Nationale Superieure des
Mines de Saint-Etienne, 2009.

J. Mockus (1988), \emph{Bayesian Approach to Global
Optimization}. Kluwer academic publishers.

T.J. Santner, B.J. Williams, and W.J. Notz (2003), \emph{The
design and analysis of computer experiments}, Springer.

M. Schonlau (1997), \emph{Computer experiments and global
optimization}, Ph.D. thesis, University of Waterloo.
}
\seealso{
\code{\link{EI}}
}
\author{
David Ginsbourger, Olivier Roustant and Victor Picheny.
}
\keyword{models}
\keyword{optimize}
