% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EI.R
\name{EI}
\alias{EI}
\alias{EI.grad}
\title{Expected Improvement Criterion and its Gradient}
\usage{
EI(
  x,
  model,
  plugin = NULL,
  type = c("UK", "SK"),
  minimization = TRUE,
  envir = NULL,
  proxy = FALSE
)

EI.grad(
  x,
  model,
  plugin = NULL,
  type = c("UK", "SK"),
  minimization = TRUE,
  envir = NULL,
  proxy = FALSE
)
}
\arguments{
\item{x}{A vector representing the input for which one wishes to
calculate EI.}

\item{model}{An object of class \code{\link[DiceKriging]{km}}.}

\item{plugin}{Optional scalar: if provided, it replaces the
minimum of the current observations.}

\item{type}{\code{"UK"} (default) or \code{"SK"}, depending
whether uncertainty related to trend estimation has to be
taken into account.}

\item{minimization}{Logical specifying if EI is used in
minimization or in maximization.}

\item{envir}{An optional environment specifying where to assign
intermediate values for future gradient calculations. Default
is \code{NULL}.}

\item{proxy}{Optional logical. If \code{TRUE}, EI is replaced by
the kriging mean, to be \emph{minimized}.}
}
\value{
The expected improvement as defined in \bold{Details}.
    If \code{plugin} is specified, its provided value will replace
    \eqn{\min Y(X)}{min Y(X)} in the formula.
}
\description{
The function \code{EI} computes the Expected
    Improvement at current location \code{x} while \code{EI.grad}
    compute the gradient at \code{x}. The current minimum of the
    observations in \code{model} can be replaced by an arbitrary
    value (plugin), which is useful in particular in noisy
    frameworks.
}
\details{
The Expected Improvement (EI) is defined as \deqn{EI(x)
    := E\left[\{\min Y(X) - Y(x)\}_+ | Y(X) = y(X) \right],}{ E[{
    min Y(X) - Y(x) }_+ | Y(X) = y(X)]} where \eqn{X} is the
    current design of experiments and \eqn{Y} is the random
    process assumed to have generated the objective function
    \eqn{y} and \eqn{z_{+} := \max\{z, \, 0)}{z_+ = max(z, 0)}
    denotes the positive part of a real number \eqn{z}. The value
    of EI is non-negative but can be numerically zero close to the
    inputs used in \code{model}. The EI and its gradient are
    computed using their closed forms.
}
\examples{
set.seed(123)
## =========================================================================
## 	EI Surface Associated with an Ordinary Kriging Model for the Branin    
##  Function Known at a 9-Points Factorial Design  
## =========================================================================

## a 9-points factorial design, and the corresponding response
## ===========================================================
d <- 2; n <- 9
design.fact <-
    expand.grid(x1 = seq(0, 1, length = 3), x2 = seq(0, 1, length = 3))
y <- apply(design.fact, 1, branin) 

## model fit
## =========
fit1 <- km(~1, design = design.fact, response = y, covtype = "gauss",
           control = list(pop.size = 50, trace = FALSE), parinit = c(0.5, 0.5))

## computing the EI
## ================
x <- c(x1 = 0.2, x2 = 0.4)
EI(x, model = fit1)
EI.grad(x, model = fit1)

## graphics
## ========
contours(object = fit1, which = character(0), grad = TRUE,
         other = "EI", otherGrad = "EI.grad",
         whereGrad = "grid", nGrid = 30) +
    ggtitle("Expected Improvement and its gradient")


}
\references{
D. Ginsbourger (2009), \emph{Multiples métamodèles pour
l'approximation et l'optimisation de fonctions numériques
multivariables}, Ph.D. thesis, Ècole Nationale Supérieure des
Mines de Saint-Ètienne.

D.R. Jones, M. Schonlau, and W.J. Welch (1998), Efficient global
optimization of expensive black-box functions, \emph{Journal of Global
Optimization}, 13, 455-492.

J. Mockus (1988), \emph{Bayesian Approach to Global
Optimization}. Kluwer academic publishers.

T.J. Santner, B.J. Williams, and W.J. Notz (2003), \emph{The
design and analysis of computer experiments}, Springer.

M. Schonlau (1997), \emph{Computer experiments and global
optimization}, Ph.D. thesis, University of Waterloo.
}
\seealso{
\code{\link{max_EI}}, \code{\link{EGO.nsteps}},
    \code{\link{qEI}}
}
\author{
David Ginsbourger, Olivier Roustant and Victor Picheny.
}
\keyword{models}
\keyword{optimize}
