% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{trend}
\alias{trend}
\title{Trend part of \code{km} object}
\usage{
trend(object, X, deriv = 0, diagOnly = FALSE)
}
\arguments{
\item{object}{A \code{km} object.}

\item{X}{A "design": a numeric matrix or data frame wih numeric
columns compatible with \code{object}.}

\item{deriv}{Logical. If \code{TRUE} the derivative will be
returned as the \code{"deriv"} attribute of the result.  This
will be an array with dimension either \code{c(n, p, n, d)} or
\code{c(n, p, d)} if \code{diag = FALSE}.}

\item{diagOnly}{Logical. If \code{TRUE} the structure of the
result takes into account the fact that the row \eqn{i} of the
trend matrix \eqn{\mathbf{F}}{F} depends only on the row
\enw{i} of the design matrix \eqn{\mathbf{X}}{X}. Then the
deivative is a 3-dimensional array with its element \eqn{i, j,
k} equal to \eqn{\partial F_{i,j}/ \partial X_{i,k}}{dF[i,
j]/dX[i, k]}.}
}
\value{

}
\description{
Evaluate the trend component of a \code{km} object on the
given design \code{X}.
}
\examples{
library(DiceKriging)
# a 16-points factorial design, and the corresponding response
d <- 2; n <- 16
design.fact <- expand.grid(x1 = seq(0, 1, length = 4),
                           x2 = seq(0, 1, length = 4))
y <- apply(design.fact, 1, branin)
# kriging model 1 : matern5_2 covariance structure, no trend, no
# nugget effect
fit <- km(~ x1 + x2, design = design.fact, response = y)
X <- matrix(runif(n = 40), ncol = 2,
            dimnames = list(NULL, c("x1", "x2")))
fitTrend <- trend(fit, X = X, deriv = 1)
fitTrend <- trend(fit, X = X[1, ], deriv = 1)
}
